% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputstreams.R
\name{getBlock}
\alias{getBlock}
\title{Grab frame block from video stream}
\usage{
getBlock(str, nFrames)
}
\arguments{
\item{str}{an input stream}

\item{nFrames}{how many frames to grab}
}
\value{
a cimg object of depth nFrames or less (if there are fewer than nFrames frames remaining in the stream)
}
\description{
getBlock grabs the next k frames as a block, meaning as a single CImg image with depth k. Note that like getFrame *this function has side effects*, so that calling it twice generally results in two different return values.
}

