% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputstreams.R
\name{goto}
\alias{goto}
\title{Jump to a certain frame in a video}
\usage{
goto(str, frame = 1)
}
\arguments{
\item{str}{an input stream}

\item{frame}{a frame number (default 1)}
}
\description{
Depending on file format and on mysterious OpenCV behaviour you may or may not succeed in jumping to the exact location you want
}

