% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputstreams.R
\name{nframes}
\alias{nframes}
\title{Total number of frames in video stream}
\usage{
nframes(str)
}
\arguments{
\item{str}{an input stream}
}
\description{
Total number of frames in video stream
}

