% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputstreams.R
\name{opencam}
\alias{opencam}
\title{Open camera device}
\usage{
opencam(device = 0)
}
\arguments{
\item{device}{device number}
}
\value{
an istream object
}
\description{
Open camera device as an input stream ('istream') object
}

