% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outputstreams.R
\name{opendisplay}
\alias{opendisplay}
\title{Open CImg display as output stream}
\usage{
opendisplay(width = 640, height = 480, title = "")
}
\arguments{
\item{width:}{width, in pixels}

\item{height:}{height, in pixels}
}
\value{
an output stream (of class "ostream")
}
\description{
CImg comes with its own little display library, and this function opens a CImg window as an output stream, meaning you'll be able to "write" frames to the stream and they'll be displayed. Useful for debugging an analysis before saving the results to a proper stream
}

