% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputstreams.R
\name{openvideo}
\alias{openvideo}
\title{Open video file as input stream}
\usage{
openvideo(fname)
}
\arguments{
\item{fname}{file name}
}
\value{
an input stream (istream)
}
\description{
Input streams are designed to be processed frame-by-frame, to avoid having to load the entire video into memory. This function opens a video file (mpeg, avi, etc.) and makes it available to R as input stream, to be queried e.g. using getFrame.
Which file formats are supported probably depends on your OS and on how OpenCV was compiled.
}
\seealso{
getFrame
}

