% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outputstreams.R
\name{openvideowriter}
\alias{openvideowriter}
\title{Open an output stream to a video file}
\usage{
openvideowriter(fname, width = 640, height = 480, frameRate = 30)
}
\arguments{
\item{fname}{file to store the input in}

\item{width}{width in pixels}

\item{height}{height in pixels}

\item{frameRate}{frame rate in frames/sec}
}
\description{
An output stream to a video file lets you save results to the hard drive frame-by-frame.
For the moment all video streams are stored in the avi format with JPEG encoding.
Use writeFrame to write a frame to the stream. Use rm(stream) to close the stream explicitly.
}
\seealso{
writeFrame
}

