% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outputstreams.R
\name{writeFrame}
\alias{writeFrame}
\title{Write a frame to an output stream}
\usage{
writeFrame(str, im)
}
\arguments{
\item{str}{}

\item{im}{}
}
\description{
If the output stream is a file, this is an actual write operation. If it's a display stream, then the frame will be displayed.
}

